+++ public/index.php
@@ -2,4 +2,6 @@
 
 require '../vendor/autoload.php';
 
+$serviceDiscovery = require '../src/ServiceDiscovery/index.php';
+
 $router = require '../src/Routes/index.php';