+++ docs/CHANGES.md
@@ -12,4 +12,17 @@
 
 
 ## v0.0.1
+- Added a `ServiceProvider` and related features
+- Additional namespaces in `composer.json`
+- Adding classless `theme.css` in `_head.php`
+- Wiring up `$serviceDiscovery` in `public/index.php`
+- Adding Attribute
+- Adding `src/Containers/ProviderInterface.php`
+- First iteration of `src/Containers/ServiceProvider.php`
+- Light maintenance on `src/Controller.php`
+- Adding `Route` attribute to `src/Controllers/HomeController.php`
+- Database configuration in `src/Models/Database.php`
+- Documentation for `src/Router.php`
+
+## v0.0.1
 ![2025-03-30](assets/Screenshot2025-03-30.png)